requirements.txt
argcomplete==1.9.3
asn1crypto==0.24.0
backcall==0.1.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
createsend==6.0.0
cryptography==2.6.1
decorator==4.4.0
Django==2.2.1
djangorestframework==3.11.0
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
httplib2==0.12.3
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
oauth2client==4.1.3
oauthlib==3.0.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
placebo==0.9.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.4.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.4.7
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.4
wsgi-request-logger==0.4.6